Revert "move unittest to integration"
This reverts commit cddfabd91632703a0b16770f497b959d2b8135cd.
